+++ src/Classifiers/NaiveBayes.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\Classifiers;
 
-use Rubix\ML\Online;
+use Rubix\ML\Elastic;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
@@ -28,7 +28,7 @@ use RuntimeException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class NaiveBayes implements Estimator, Online, Probabilistic, Persistable
+class NaiveBayes implements Estimator, Elastic, Probabilistic, Persistable
 {
     const LOG_EPSILON = -8;