+++ src/Regressors/Adaline.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\Regressors;
 
-use Rubix\ML\Online;
+use Rubix\ML\Elastic;
 use Rubix\ML\Estimator;
 use Rubix\Tensor\Matrix;
 use Rubix\ML\Persistable;
@@ -34,7 +34,7 @@ use RuntimeException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class Adaline implements Estimator, Online, Persistable
+class Adaline implements Estimator, Elastic, Persistable
 {
     /**
      * The maximum number of training epochs. i.e. the number of times to iterate