+++ tests/Classifiers/LogisticRegressionTest.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\Tests\Classifiers;
 
-use Rubix\ML\Online;
+use Rubix\ML\Elastic;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
@@ -37,7 +37,7 @@ class LogisticRegressionTest extends TestCase
     {
         $this->assertInstanceOf(LogisticRegression::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
-        $this->assertInstanceOf(Online::class, $this->estimator);
+        $this->assertInstanceOf(Elastic::class, $this->estimator);
         $this->assertInstanceOf(Probabilistic::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
     }