+++ tests/Clusterers/KMeansTest.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\Tests\Clusterers;
 
-use Rubix\ML\Online;
+use Rubix\ML\Elastic;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Labeled;
@@ -29,7 +29,7 @@ class KMeansTest extends TestCase
     public function test_build_clusterer()
     {
         $this->assertInstanceOf(KMeans::class, $this->estimator);
-        $this->assertInstanceOf(Online::class, $this->estimator);
+        $this->assertInstanceOf(Elastic::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }