+++ tests/Regressors/MLPRegressorTest.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\Tests\Regressors;
 
-use Rubix\ML\Online;
+use Rubix\ML\Elastic;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Labeled;
@@ -43,7 +43,7 @@ class MLPRegressorTest extends TestCase
     public function test_build_regressor()
     {
         $this->assertInstanceOf(MLPRegressor::class, $this->estimator);
-        $this->assertInstanceOf(Online::class, $this->estimator);
+        $this->assertInstanceOf(Elastic::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }